split sol and evm creation
